checkpoint

+++ fixedpt.h
@@ -12,20 +12,20 @@ namespace fixed {
 
 class Num {
   static unsigned _negExponent;
-  static __int128 _scale;
+  static long long _scale;
   static Num *_lowest;
   static Num *_min;
   static Num *_max;
-  __int128 _mantissa;
-  static __int128 mantissa(__int128 value) { return value * _scale; }
-  static __int128 mantissa(int value) { return value * _scale; }
-  static __int128 mantissa(long double value) {
+  long long _mantissa;
+  static long long mantissa(long long value) { return value * _scale; }
+  static long long mantissa(int value) { return value * _scale; }
+  static long long mantissa(long double value) {
     return llrintl(value * powl(10, _negExponent));
   }
-  static __int128 mantissa(double value) {
+  static long long mantissa(double value) {
     return llrint(value * pow(10, _negExponent));
   }
-  static __int128 parse(const std::string &s);
+  static long long parse(const std::string &s);
 
  public:
   static void init(unsigned negExponent);
@@ -39,12 +39,12 @@ class Num {
   Num(double value) : _mantissa(mantissa(value)) { assert(_negExponent); }
   Num(const std::string &s) : _mantissa(parse(s)) { assert(_negExponent); }
   operator std::string() const;
-  // friend Num operator*(__int128 a, const Num &b);
+  // friend Num operator*(long long a, const Num &b);
   Num &operator+=(const Num &b) {
     _mantissa += b._mantissa;
     return *this;
   };
-  Num &operator+=(__int128 b) {
+  Num &operator+=(long long b) {
     _mantissa += mantissa(b);
     return *this;
   };
@@ -52,7 +52,7 @@ class Num {
     _mantissa -= b._mantissa;
     return *this;
   };
-  Num &operator-=(__int128 b) {
+  Num &operator-=(long long b) {
     _mantissa -= mantissa(b);
     return *this;
   };
@@ -60,7 +60,7 @@ class Num {
     _mantissa = _mantissa * b._mantissa / _scale;
     return *this;
   };
-  Num &operator*=(__int128 b) {
+  Num &operator*=(long long b) {
     _mantissa *= b;
     return *this;
   };
@@ -68,7 +68,7 @@ class Num {
     _mantissa *= b;
     return *this;
   };
-  Num &operator/=(__int128 b) {
+  Num &operator/=(long long b) {
     _mantissa /= b;
     return *this;
   };
@@ -82,7 +82,7 @@ class Num {
   }
   Num operator-() { return Num(*this).negate(); }
 
-  bool operator>(__int128 b) const { return _mantissa > mantissa(b); }
+  bool operator>(long long b) const { return _mantissa > mantissa(b); }
 };
 template <typename T>
 inline Num operator+(const Num &a, const T &b) {
